Theorycrafting for Vectrum.c
Took some time to build out a better conceptual understanding of what it
is I'm building, rather than follow some code-along tutorial. Build out
a general idea of what the data structure for a Vectrum looks like, what
the general functionality and contract might be, etc.. All noted within
a comment block in Vectrum.h. Not sure if the header file is the right
place for that, but it felt right.